VertexArray: Added Buffers Getters

+++ TituEngine/src/TituEngine/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -37,7 +37,7 @@ namespace TituEngine
 		for (const auto& element : layout)
 		{
 			glEnableVertexAttribArray(index);
-			glVertexAttribPointer(index, element.ElementCount, GL_FLOAT, element.Normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (const void*)element.Offset);
+			glVertexAttribPointer(index, element.ElementCount, GL_FLOAT, element.Normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (const void*)(intptr_t)element.Offset);
 			index++;
 		}
